+++ CHANGELOG.md
@@ -87,6 +87,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#3673](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3673))
 - `opentelemetry-instrumentation-starlette`/`opentelemetry-instrumentation-fastapi`: Fixes a crash when host-based routing is used
   ([#3507](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3507))
+- `opentelemetry-instrumentation-django`: Fixes invalid type at WSGI request headers and attributes collection.
+  ([#3731](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3731))
 - Fix documentation order of sections and headers for Django, Flask, MySQL, mysqlclient, psycopg, psycopg2, pymysql, sqlalchemy instrumentations.
   ([#3719](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/3719))
 - `opentelemetry-instrumentation-asgi` Fixed an issue where FastAPI reports IP instead of URL.